Add test for default arg value's bytecode generation.
